Add files via upload
This code sets up a basic list with five items in HTML. The CSS code sets the default background color of the list items to white and the default color to black. The transition property is used to smoothly change the color when it's updated in JavaScript.

The JavaScript code uses setInterval() to repeatedly update the color of each list item every second. The getRandomColor() function generates a random hex color code, which is then used to set the color of each list item.